requirements.txt
Django
django-rest-framework
numpy
pandas
sklearn
scipy
csv2json
